FWMT-554 made multiplier double to fix miss match exception

+++ src/main/java/uk/gov/ons/fwmt/job_service_v2/config/QueueConfig.java
@@ -30,11 +30,11 @@ import static uk.gov.ons.fwmt.fwmtgatewaycommon.config.QueueNames.JOB_SVC_ADAPTE
 public class QueueConfig {
 
   private int initialInterval;
-  private int multiplier;
+  private double multiplier;
   private int maxInterval;
 
   public QueueConfig(@Value("${rabbitmq.initialinterval}") Integer initialInterval,
-      @Value("${rabbitmq.multiplier}") Integer multiplier,
+      @Value("${rabbitmq.multiplier}") Double multiplier,
       @Value("${rabbitmq.maxInterval}") Integer maxInterval) {
     this.initialInterval = initialInterval;
     this.multiplier = multiplier;
